refactor(slotting): Phase 4 slice 7b — the slotting Polars namespace becomes plain typed functions
engine/slotting/namespace.py (469 LOC, SlottingLazyFrame + SlottingExpr)
is deleted; engine/slotting/transforms.py lands the 9 public transforms
with bodies verbatim (prepare_columns, apply_slotting_weights @cites
Art. 153(5), calculate_rwa, apply_el_rates, compute_el_shortfall_excess,
apply_all, build_audit, lookup_rw, lookup_el_rate), composed via .pipe at
all call sites: 5 in SlottingCalculator.calculate_branch, 107 across 5
test files (test_slotting_namespace.py renamed to
test_slotting_transforms.py). _SHORT_MATURITY_THRESHOLD_YEARS moved with
its consumer; allowlists re-keyed; citation snapshot re-homed (1 key);
zero residual namespace references.

Conversion template for the sa/irb retirements. Gates: unit+contracts
6,065 (pre-change minus exactly the deliberate registration-class
rewrite); acceptance+integration 1,482; parity byte-identical;
ruff/ty/arch_check green.
