+++ tests/contracts/data/citation_snapshot.json
@@ -380,7 +380,7 @@
   "rwa_calc.engine.slotting.calculator::SlottingCalculator.calculate_branch": [
     "CRR Art. 153(5)"
   ],
-  "rwa_calc.engine.slotting.namespace::SlottingLazyFrame.apply_slotting_weights": [
+  "rwa_calc.engine.slotting.transforms::apply_slotting_weights": [
     "CRR Art. 153(5)"
   ],
   "rwa_calc.engine.stages.classify.approach::_align_irb_exposure_class": [
